+++ src/Geometry/Octree.cpp
@@ -134,7 +134,7 @@ void Octree::Initialize(int min_scale_level, int max_scale_level) {
 	root.Clear();
 	root.level = this->max_scale_level;
 	
-	max_len = FastPow<double>(2.0, this->max_scale_level);
+	max_len = FastPow<float>(2.0f, (float)this->max_scale_level);
 	max_off = max_len / 2;
 	root.position = vec3(-max_off, -max_off, -max_off);
 	
@@ -222,7 +222,7 @@ uint64 Octree::GetSeekBits(vec3 pos, int level) const {
 	uint64 seek = 0, j = 0;
 	for(int shift = rev_level_i - 1; shift >= 0; shift--) {
 		uint32 mask = 1UL << shift;
-		for(uint64 i = 0; i < 3; i++)
+		for(int i = 0; i < 3; i++)
 			if (ipos[i] & mask)
 				seek |= 1ULL << (j + i);
 		j += 3;
